+++ graphbrain/tests/test_hyperedge.py
@@ -337,10 +337,13 @@ class TestHyperedge(unittest.TestCase):
     def test_atom_type(self):
         self.assertEqual(hedge('graphbrain/Cp.s/1').type(), 'Cp')
 
+    def test_atom_mtype(self):
+        self.assertEqual(hedge('graphbrain/Cp.s/1').mtype(), 'C')
+
     def test_atom_type_implied_conjunction(self):
         self.assertEqual(hedge('and').type(), 'J')
 
-    def test_entity_type(self):
+    def test_non_atom_type(self):
         self.assertEqual(hedge('(is/Pd.so graphbrain/Cp.s great/C)').type(),
                          'Rd')
         self.assertEqual(hedge('(red/M shoes/Cc.p)').type(), 'Cc')
@@ -354,6 +357,21 @@ class TestHyperedge(unittest.TestCase):
         self.assertEqual(hedge('(and/J meat/Cc.s potatoes/Cc.p)').type(), 'C')
         self.assertEqual(
             hedge('(and/J (is/Pd.so graphbrain/Cp.s great/C))').type(), 'R')
+
+    def test_non_atom_mtype(self):
+        self.assertEqual(hedge('(is/Pd.so graphbrain/Cp.s great/C)').type(),
+                         'Rd')
+        self.assertEqual(hedge('(red/M shoes/Cc.p)').mtype(), 'C')
+        self.assertEqual(hedge('(before/Tt noon/C)').mtype(), 'S')
+        self.assertEqual(hedge('(very/M large/M)').mtype(), 'M')
+        self.assertEqual(hedge('((very/M large/M) shoes/Cc.p)').mtype(), 'C')
+        self.assertEqual(hedge('(will/M be/Pd.sc)').mtype(), 'P')
+        self.assertEqual(hedge('((will/M be/Pd.sc) john/Cp.s rich/C)').mtype(),
+                         'R')
+        self.assertEqual(hedge('(play/T piano/Cc.s)').mtype(), 'S')
+        self.assertEqual(hedge('(and/J meat/Cc.s potatoes/Cc.p)').mtype(), 'C')
+        self.assertEqual(
+            hedge('(and/J (is/Pd.so graphbrain/Cp.s great/C))').mtype(), 'R')
 
     def test_connector_type(self):
         self.assertEqual(hedge('graphbrain/Cp.s/1').connector_type(), None)
@@ -369,6 +387,21 @@ class TestHyperedge(unittest.TestCase):
         self.assertEqual(hedge('((will/M be/Pd.sc) john/Cp.s '
                                'rich/C)').connector_type(), 'Pd')
         self.assertEqual(hedge('(play/T piano/Cn.s)').connector_type(), 'T')
+
+    def test_mconnector_type(self):
+        self.assertEqual(hedge('graphbrain/Cp.s/1').mconnector_type(), None)
+        self.assertEqual(hedge('graphbrain').mconnector_type(), None)
+        self.assertEqual(hedge('(is/Pd.so graphbrain/Cp.s '
+                               'great/C)').mconnector_type(), 'P')
+        self.assertEqual(hedge('(red/M shoes/Cn.p)').mconnector_type(), 'M')
+        self.assertEqual(hedge('(before/Tt noon/C)').mconnector_type(), 'T')
+        self.assertEqual(hedge('(very/M large/M)').mconnector_type(), 'M')
+        self.assertEqual(hedge('((very/M large/M) '
+                               'shoes/Cn.p)').mconnector_type(), 'M')
+        self.assertEqual(hedge('(will/M be/Pd.sc)').mconnector_type(), 'M')
+        self.assertEqual(hedge('((will/M be/Pd.sc) john/Cp.s '
+                               'rich/C)').mconnector_type(), 'P')
+        self.assertEqual(hedge('(play/T piano/Cn.s)').mconnector_type(), 'T')
 
     def test_atom_with_type(self):
         self.assertEqual(hedge('(+/B a/Cn b/Cp)').atom_with_type('C'),